commit
commit

+++ routes/ideas.js
@@ -97,19 +97,31 @@ router.put("/:id", async (req, res) => {
 
   // res.json({ success: true, data: idea });
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      {
-        new: true,
-      }
-    );
-    res.json({ success: true, data: updatedIdea });
+        {
+          new: true,
+        }
+      );
+
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: "You are not authorized to update this resource.",
+    });
   } catch (error) {
     res.status(500).json({ success: false, error: "Something went wrong." });
   }
@@ -130,8 +142,19 @@ router.delete("/:id", async (req, res) => {
 
   // res.json({ success: true, data: {} });
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: "You are not authorized to delete this resource.",
+    });
   } catch (error) {
     res.status(500).json({ success: false, error: "Something went wrong." });
   }
